test: Eliminate warnings with GCC 13
../test/dtor_queue_dso.c: In function ‘main’:
../test/dtor_queue_dso.c:24:31: warning: unknown conversion type character ‘t’ in format [-Wformat=]
   24 |       fprintf(stderr, "push: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
      |                               ^
../test/dtor_queue_dso.c:24:43: warning: format ‘%d’ expects argument of type ‘int’, but argument 3 has type ‘long long int’ [-Wformat=]
   24 |       fprintf(stderr, "push: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
      |                                          ~^     ~~~~~~~~~~~~~~~~~~~~~~
      |                                           |     |
      |                                           int   long long int
      |                                          %lld

+++ test/dtor_queue_all.c
@@ -21,13 +21,13 @@ main(void)
 
       r = __MCF_dtor_queue_push(&queue, &elem);
       assert(r == 0);
-      fprintf(stderr, "push: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
+      fprintf(stderr, "push: %d, size = %d\n", (int)(intptr_t) elem.__dtor, queue.__size);
     }
 
     for(intptr_t k = 1000;  k >= 1;  --k) {
       r = __MCF_dtor_queue_pop(&elem, &queue, NULL);
       assert(r == 0);
-      fprintf(stderr, "pop: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
+      fprintf(stderr, "pop: %d, size = %d\n", (int)(intptr_t) elem.__dtor, queue.__size);
 
       assert(elem.__dtor == (__MCF_cxa_dtor_cdecl*) k);
       assert(elem.__dso == (void*) (k % 10));
