+++ test/dtor_queue_dso.c
@@ -21,7 +21,7 @@ main(void)
 
       r = __MCF_dtor_queue_push(&queue, &elem);
       assert(r == 0);
-      fprintf(stderr, "push: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
+      fprintf(stderr, "push: %d, size = %d\n", (int)(intptr_t) elem.__dtor, queue.__size);
 
     }
 
@@ -29,7 +29,7 @@ main(void)
       for(intptr_t k = 99;  k >= 0;  --k) {
         r = __MCF_dtor_queue_pop(&elem, &queue, (void*) m);
         assert(r == 0);
-        fprintf(stderr, "pop: %td, size = %d\n", (intptr_t) elem.__dtor, queue.__size);
+        fprintf(stderr, "pop: %d, size = %d\n", (int)(intptr_t) elem.__dtor, queue.__size);
 
         assert(elem.__dtor == (void*) (k * 10 + m));
         assert(elem.__dso == (void*) m);
